backend/requirements.txt
# Web Scraping
scrapy==2.11.0
scrapy-playwright==0.0.34
playwright==1.40.0
scrapy-user-agents==0.1.1
scrapy-rotating-proxies==0.6.2

# Data Processing & NLP
spacy==3.7.2
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
nltk==3.8.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Databases
elasticsearch==8.11.0
neo4j==5.15.0
pymongo==4.6.0
redis==5.0.1

# API Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Science & Visualization
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
pyyaml==6.0.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
